requirements_external.txt
# Requirements for external deployment (Render.com, Heroku, etc.)
python-telegram-bot>=20.0
google-genai>=0.7.0
together>=1.0.0
flask>=2.3.0